add new compute shader example: clifford-chaos-game
